backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain-community
langchain-openai
langchain-pinecone
langchain-text-splitters
langchain
chromadb
openai
tiktoken
